Moves Telecom API to mainline module
Moves the Telecom API to a mainline module and defines
the framework-telecom target as per flagging recommendations.

Note this CL simply copies the API into a module. The next dependent
CL switches the build system to remove the API from the FW and
use the module API.

Bug: 400505269
Bug: 405460896
Test: build + test phone calling
Flag:  build.RELEASE_TELECOM_MAINLINE_MODULE
Change-Id: Id4ad467c492e2854e2cef81fc23c6c8007b1d758