Add new flag lib for only internal flags
Adds a new flag lib for internal flags because
we could not use the existing android.telecom.flags
flag lib for non-exported flags without creating a
runtime exception.

The cause of the runtime exception is not clear, but
it seems to be a side effect of using the default
"framework-telecom-non-updatable-defaults" when the telecom
release flag is off.

Test: build test flag and ensure TelecomLib can access
Flag: build.RELEASE_TELECOM_MAINLINE_MODULE
Bug: 445218695
Change-Id: I954c678e9137d42a36d66db7ec57d53298042f6b